[refactor🔨] : 회원번호를 Long에서 Integer로 변경

+++ src/main/java/shop/gaship/gashipfront/security/social/dto/domain/Member.java
@@ -13,7 +13,7 @@ import lombok.Setter;
 @Setter
 @EqualsAndHashCode
 public class Member {
-    private Long identifyNo;
+    private Integer identifyNo;
     private String email;
     private List<String> authorities;
     private String password;
